marketAnalysis/requirements.txt
# Core dependencies for Meme Coin Market Analysis Model

# Data manipulation and analysis
pandas>=1.5.0
numpy>=1.21.0

# API requests and environment management
requests>=2.28.0
python-dotenv>=0.19.0

# Technical analysis indicators
ta>=0.10.0

# Machine learning and deep learning
scikit-learn>=1.1.0
tensorflow>=2.10.0
joblib>=1.2.0

# Data visualization (optional but recommended)
matplotlib>=3.5.0
seaborn>=0.11.0

# Development and testing (optional)
pytest>=7.0.0
jupyter>=1.0.0
